Checkpoint. Can create trigger and source together
